requirements.txt
fastapi
uvicorn
sqlmodel
google-genai
python-dotenv
pydantic-settings
requests
httpx
alembic
psycopg2-binary
numpy
pgvector
slowapi
celery
redis
